requirements.txt
fastapi
uvicorn
pymongo
python-dotenv
passlib[bcrypt]
pydantic